+++ packages/article-label/__tests__/ios/jest.config.js
@@ -1,5 +1,3 @@
-const jestConfigurator = require("@times-components/jest-configurator");
+const jestConfigurator = require("@times-components/jest-configurator").default;
 
-module.exports = Object.assign(jestConfigurator("article-label", "ios"), {
-  collectCoverageFrom: ["**/packages/article-label/article-label.js"]
-});
+module.exports = jestConfigurator("article-label", "ios");