+++ packages/link/__tests__/web/jest.config.js
@@ -1,12 +1,3 @@
-const jestConfigurator = require("@times-components/jest-configurator");
+const jestConfigurator = require("@times-components/jest-configurator").default;
 
-module.exports = Object.assign(jestConfigurator("link", "web"), {
-  collectCoverageFrom: [
-    "**/packages/link/**/*.js",
-    "!**/packages/link/**/link.js",
-    "!**/packages/link/**/link.android.js",
-    "!**/packages/link/**/*stories*",
-    "!**/packages/link/__tests__/**",
-    "!**/packages/link/coverage/**"
-  ]
-});
+module.exports = jestConfigurator("link", "web");