+++ packages/watermark/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.2.7"></a>
+## [0.2.7](https://github.com/newsuk/times-components/compare/@times-components/watermark@0.2.6...@times-components/watermark@0.2.7) (2018-01-08)
+
+
+
+
+**Note:** Version bump only for package @times-components/watermark
+
+<a name="0.2.6"></a>
+## [0.2.6](https://github.com/newsuk/times-components/compare/@times-components/watermark@0.2.5...@times-components/watermark@0.2.6) (2018-01-05)
+
+
+
+
+**Note:** Version bump only for package @times-components/watermark
+
 <a name="0.2.5"></a>
 ## [0.2.5](https://github.com/newsuk/times-components/compare/@times-components/watermark@0.2.4...@times-components/watermark@0.2.5) (2018-01-02)